config-generator/pyproject.toml
[project]
name = "config-generator"
version = "0.1.0"
description = "AI-powered config generator using Qwen 7B"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "rich>=13.7.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/generator"]

[project.scripts]
generate-config = "generator.cli:main"